requirements.txt
alembic==1.16.1
azure-core==1.34.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.4
databricks-cli==0.18.0
docker==7.1.0
entrypoints==0.4
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
gunicorn==23.0.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lockfile==0.12.2
Mako==1.3.10
MarkupSafe==3.0.2
mlflow==1.26.1
model-helpers==0.1.0
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
packaging==25.0
pandas==2.3.0
pid==3.0.4
pillow==11.2.1
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
protobuf==3.20.0
pycparser==2.22
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
service==0.6.0
setproctitle==1.3.6
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.23.0
